Fixes for unbound lambdas
Unbound lambdas are lambdas that are not found directly on the right side of a let binding. Since the compiler cannot handle these they are translated to core as let bindings which directly loads the bound lambda which causes the lambda lifter to lift them out of the expression entirely. The compiler was changed to fail if a lambda is found anywhere else than directly in the binding