Made typecheck return a Type which is used when typechecking
To avoid having to worry about storing the correct type at each expression we start to move towards only working with the type returned from the typecheck function. When the typechecking is complete we should still be able to get all types since all needed types are bound to a name.