Translation of nested patterns into case expressions
This passes all current tests but is still a work in progress. The basic idea is to first pull out all nested patterns into a flat structure. After that we walk through this list and group all patterns together which matches the same thing and create an alternative which matches this and we generate more nested case expressions with the rest of those pattern lists that matches.