Changed type variables to internally hold InternedStr
InternedStr helps with error messages since the actual variables that are declared can be used.

Since the int that were originally used were also used to determine to determine which variable was created first an 'age' field were added to preserve this usage. This field should be possible to remove with some changes to the constraints handling (notable when adding other Assemblys/Types trait objects throgh add_types)