Fixed typevariable collisions when typechecking with multiple TypeEnvironments
Previously, doing the typechecking with multiple environments could cause collisions. For instance, when typechecking a module the function sum could get the type 'sum :: Num 3 => [3]' which when later on another environment was created the variable 3 would be reused and with 'sum' imported through the 'add_types' function the variable 3 would already have the 'Num' constraint causing the typechecker to fail if this was not expected. By storing the constraints with the type we can add the constraints to the environment as the imported binding is instantiated with fresh variables.