Rough implementation of runtime construction of dictionaries
The dictionary type for the vm was extended to allow it not only hold function indices but also functions which hold their own dictionary.
Example:
instance Eq a => Eq [a] ...
When loading the dictionary for this instance each function in the instance needs to have the dictionary for Eq a applied to it. This forces the vm to do this at runtime.
The current way of doing this is very inefficient but it will do until more testing can be done.