Changed the typecheck error handling to store all errors instead of failing
Currently typecheck still calls fail! if any errors are found since many tests depend on it failing if an error is found.
This way of handling errors may still need to be tuned since after an error is found, typechecking still continues which could possible cause a chain reaction causing more errors than necessary to be reported. For now this seems like an improvement.