Global names have no longer always uid == 0
uid == 0 is reserved for the Prelude to allow code generation and translation to have a simple construction of True, False etc.

For all other modules a single uid is defined for the global bindings in that module.
All instances and classes use the same uid as the one defined for the class in question. This is not fully implemented in the Renamer as of now